backend/pyproject.toml
[tool.poetry]
name = "slo-scout-backend"
version = "1.0.0"
description = "SLO-Scout backend API and control plane"
authors = ["SLO-Scout Team"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
kafka-python = "^2.0.2"
pymilvus = "^2.3.0"
boto3 = "^1.28.0"
sentence-transformers = "^2.2.0"
openai = "^1.0.0"
onnxruntime = "^1.16.0"
prometheus-client = "^0.18.0"
httpx = "^0.25.0"
tenacity = "^8.2.3"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.10.0"
ruff = "^0.1.0"
mypy = "^1.6.0"
httpx = "^0.25.0"
testcontainers = "^3.7.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
